+++ datafusion/core/src/physical_plan/aggregates/mod.rs
@@ -45,6 +45,7 @@ use std::any::Any;
 use std::collections::HashMap;
 use std::sync::Arc;
 
+mod group_values;
 mod no_grouping;
 mod order;
 mod row_hash;
